#3 Coupled cobra with viper and added "verbose" flag
With cobra and viper coupled, we can now override configuration via
command line arguments and environment variables.

A persistent flag is already set "verbose" and illustrate the point.
It enables to see debug logs.

Our idea at first was to create an helper which would automatically
create a new flag, bind it to the corresponding command and config file
key. However, due to the way spf13/viper and spf13/pflags are made, I
didn't found a way of doing this. Indeed, creating a flag requires a
Value and the standard values factories are private in the pflag library.

This means the author wants us to pass via the flag.<TYPE>(....) which
register a flag automatically and bind the output variable to the
reference of the flag value receiver. We have no trace of the flag just
created which is handled internally.

Because pops will be a medium size project and we have more important
features to implement, I give up. At least, until this will be really needed.